feat(node): pluggable logging class (support different strategies)
