+++ packages/node/src/lib/construct-payload.ts
@@ -1,3 +1,4 @@
+import type { LoggerStrategy } from './logger';
 import type { OutgoingLogBody } from './metrics-log';
 import type { UUID } from 'node:crypto';
 import type { IncomingMessage, ServerResponse } from 'node:http';
@@ -55,9 +56,9 @@ export interface LogOptions {
   fireAndForget?: boolean;
 
   /**
-   * If true, the errors and other logs will be displayed in console.
+   * If true, the errors and other logs will be displayed in console. Your own logger strategy can be passed.
    */
-  logger?: boolean;
+  logger?: LoggerStrategy | boolean;
 
   /**
    * @deprecated use `allowList` instead
